Parallel loader queue clean up
Jiting an image that gets removed from the DOM would eventualy fill up the queue and would block the loader indefinitly.The change made here is to clean up the queue from dead elements before the processing the queue. We also make sure to decrement the active count to be able to continue further processing.